% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konverter.R
\name{adata2SCE}
\alias{adata2SCE}
\title{AnnData to SingleCellExperiment}
\usage{
adata2SCE(adata)
}
\arguments{
\item{adata}{Reference to a Python AnnData object}
}
\value{
SingleCellExperiment
}
\description{
Converts a Python AnnData object to a SingleCellExperiment object.
}
