% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konverter.R
\name{SCE2AnnData}
\alias{SCE2AnnData}
\title{SingleCellExperiment to AnnData}
\usage{
SCE2AnnData(sce, X_name = NULL)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{X_name}{Name of the assay to use as the X matrix of the AnnData object.
If \code{NULL} the first assay will be used.}
}
\value{
Reference to a Python AnnData object
}
\description{
Converts a SingleCellExperiment object a Python AnnData object.
}
