% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{setZellkonverterVerbose}
\alias{setZellkonverterVerbose}
\title{Set zellkonverter verbose}
\usage{
setZellkonverterVerbose(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical value for the verbosity option.}
}
\value{
The value of getOption("zellkonverter.verbose") invisibly
}
\description{
Set the zellkonverter verbosity option
}
\details{
Running \code{setZellkonverterVerbose(TRUE)} will turn on \strong{zellkonverter}
progress messages by default without having to set \code{verbose = TRUE} in each
function call. This is done by setting the \code{"zellkonverter.verbose"} option.
Running \code{setZellkonverterVerbose(FALSE)} will turn default verbosity off.
}
\examples{
current <- getOption("zellkonverter.verbose")
setZellkonverterVerbose(TRUE)
getOption("zellkonverter.verbose")
setZellkonverterVerbose(FALSE)
getOption("zellkonverter.verbose")
setZellkonverterVerbose(current)
getOption("zellkonverter.verbose")
}
