% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zellkonverter-package.R
\docType{package}
\name{zellkonverter-package}
\alias{zellkonverter}
\alias{zellkonverter-package}
\title{zellkonverter: Conversion Between scRNA-seq Objects}
\description{
Provides methods to convert between Python AnnData objects and SingleCellExperiment objects. These are primarily intended for use by downstream Bioconductor packages that wrap Python methods for single-cell data analysis. It also includes functions to read and write H5AD files used for saving AnnData objects to disk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/theislab/zellkonverter}
  \item Report bugs at \url{https://github.com/theislab/zellkonverter/issues}
}

}
\author{
\strong{Maintainer}: Luke Zappia \email{luke@lazappi.id.au} (\href{https://orcid.org/0000-0001-7744-8565}{ORCID}) (lazappi)

Authors:
\itemize{
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID}) (LTLA)
}

Other contributors:
\itemize{
  \item Jack Kamm \email{jackkamm@gmail.com} (\href{https://orcid.org/0000-0003-2412-756X}{ORCID}) (jackkamm) [contributor]
  \item Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood) [contributor]
  \item Gabriel Hoffman \email{gabriel.hoffman@mssm.edu} (\href{https://orcid.org/0000-0002-0957-0224}{ORCID}) (GabrielHoffman) [contributor]
  \item Marek Cmero \email{cmero.ma@wehi.edu.au} (\href{https://orcid.org/0000-0001-7783-5530}{ORCID}) (mcmero) [contributor]
}

}
