% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{AnnData-Environment}
\alias{AnnData-Environment}
\alias{.AnnDataDependencies}
\alias{zellkonverterAnnDataEnv}
\title{AnnData environment}
\format{
A character vector containing the pinned versions of all Python packages in
\code{zellkonverterAnnDataEnv}.

A \code{\link[basilisk:BasiliskEnvironment-class]{basilisk::BasiliskEnvironment()}} containing \strong{zellkonverter}'s AnnData
Python environment.
}
\usage{
.AnnDataDependencies

zellkonverterAnnDataEnv
}
\description{
The Python environment used by \strong{zellkonverter} for interfacing with the
\strong{anndata} Python library (and H5AD files) is described by the dependencies
in \code{.AnnDataDependencies}. The \code{zellkonverterAnnDataEnv} variable is the
\code{\link[basilisk:BasiliskEnvironment-class]{basilisk::BasiliskEnvironment()}} used by \strong{zellkonverter}.
}
\details{
The \code{.AnnDataDependencies} variable is exposed for use by other package
developers who want an easy way to define the dependencies required for
creating a Python environment to work with AnnData objects, most typically
within a \strong{basilisk} context. For example, we can simply combine this
vector with additional dependencies to create a \strong{basilisk} environment with
Python package versions that are consistent with those in \strong{zellkonverter}.

If you want to run code in the exact environment used by \strong{zellkonverter}
this can be done using \code{zellkonverterAnnDataEnv} in combination with
\code{\link[basilisk:basiliskStart]{basilisk::basiliskStart()}} and/or \code{\link[basilisk:basiliskStart]{basilisk::basiliskRun()}}. Please refer to
the \strong{basilisk} documentation for more information on using these
environments.
}
\examples{
.AnnDataDependencies

cl <- basilisk::basiliskStart(zellkonverterAnnDataEnv)
anndata <- reticulate::import("anndata")
basilisk::basiliskStop(cl)

}
\author{
Luke Zappia

Aaron Lun
}
\keyword{datasets}
