% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{AnnDataDependencies}
\alias{AnnDataDependencies}
\alias{.AnnDataDependencies}
\title{AnnData dependencies}
\format{
A character vector containing the pinned versions of all Python packages on
which AnnData depends.
}
\usage{
.AnnDataDependencies
}
\description{
Vector defining a set of Python dependencies and versions required to operate
with AnnData and H5AD files.
}
\details{
This variable is exposed for use by other package developers who want an easy
way to define the dependencies required for creating a Python environment to
work with AnnData objects, most typically within a \strong{basilisk} context. For
example, we can simply combine this vector with additional dependencies to
create a \strong{basilisk} environment with Python package versions that are
consistent with those in \strong{zellkonverter}.
}
\examples{
.AnnDataDependencies

}
\author{
Luke Zappia

Aaron Lun
}
\keyword{datasets}
