% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readH5AD}
\alias{readH5AD}
\title{Read H5AD}
\usage{
readH5AD(file)
}
\arguments{
\item{file}{Path to a \code{.h5ad} file}
}
\value{
SingleCellExperiment
}
\description{
Reads a H5AD file and returns a SingleCellExperiment
}
