% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readH5AD}
\alias{readH5AD}
\title{Read H5AD}
\usage{
readH5AD(file)
}
\arguments{
\item{file}{String containing a path to a \code{.h5ad} file.}
}
\value{
A \linkS4class{SingleCellExperiment} object is returned.
}
\description{
Reads a H5AD file and returns a \linkS4class{SingleCellExperiment} object.
}
\details{
When first run, this function will instantiate a conda environment
containing all of the necessary dependencies.
This will not be performed on any subsequent run or if any other
\pkg{zellkonverter} function has been run prior to this one.
}
\examples{
# Re-using the example from writeH5AD.
example(writeH5AD, echo=FALSE)

# Reading into a SingleCellExperiment.
sce2 <- readH5AD(temp)

}
\seealso{
\code{\link{writeH5AD}}, to write a SingleCellExperiment file to a H5AD file.

\code{\link{AnnData2SCE}}, for developers to convert existing AnnData instances to a SingleCellExperiment.
}
\author{
Luke Zappia
}
