% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeH5AD}
\alias{writeH5AD}
\title{Write H5AD}
\usage{
writeH5AD(sce, file)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{file}{Path to write new \code{.h5ad}}
}
\value{
SingleCellExperiment
}
\description{
Write a H5AD file from a SingleCellExperiment
}
