% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zellkonverter-package.R
\docType{package}
\name{zellkonverter-package}
\alias{zellkonverter}
\alias{zellkonverter-package}
\title{zellkonverter: Conversion between scRNA-seq objects}
\description{
Zellkonverter is a small package for converting between
    SingleCellExperiment objects and alternative objects for storing single-cell
    RNA-sequencing data (such as AnnData).
}
\author{
\strong{Maintainer}: Luke Zappia \email{luke@lazappi.id.au} (\href{https://orcid.org/0000-0001-7744-8565}{ORCID})

}
\keyword{internal}
