% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zellkonverter-package.R
\docType{package}
\name{zellkonverter-package}
\alias{zellkonverter}
\alias{zellkonverter-package}
\title{zellkonverter: Conversion between scRNA-seq objects}
\description{
Provides methods to convert AnnData Python objects to and from
    SingleCellExperiment objects, as well as functions to read/write
    H5AD files. This is intended for use by downstream Bioconductor 
    packages that wrap Python methods for single-cell data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/theislab/zellkonverter}
  \item Report bugs at \url{https://github.com/theislab/zellkonverter/issues}
}

}
\author{
\strong{Maintainer}: Luke Zappia \email{luke@lazappi.id.au} (\href{https://orcid.org/0000-0001-7744-8565}{ORCID})

Authors:
\itemize{
  \item Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com} (\href{https://orcid.org/0000-0002-3564-4813}{ORCID})
}

}
\keyword{internal}
